Moved .flex container .content-left to responsive section and added additional comments
Moved the .flex-container .content-left / content-right to the responsive section, under desktop.

This makes it so that, in desktop mode, e.g. homepage, there are two columns, blog on the left, pictures on right.

Then, outside of desktop mode (mobile), the flex box will cause the containers to form a column instead, also making the blog in mobile be wider than before, more readable.